Action classes renamed to ResourceAction classes.
